Add a more thorough treatment of statelessness
This is more than @toreini added in #26, but I realize that this is an
important point and one that can be subject to technical measures.
The real challenge here is governance of any such system, but it is
possible to remove choice when it comes to the use of digital
credentials around certain traits.  If the point of the credential is to
do things like establish basic traits, like name, birthdate, or that
this is a human being, it should not matter who makes this assertion.

Closes #23.  Though I did not expect to be fixing that, this seems to
have worked out.