Write something about wallets (#31)
* Write something about wallets

I wrote something very similar to this in another context today and it
seemed like it might help here.  It's not that we don't want to trust
that wallets will be effective advocates for protecting the interests of
their users, it's that we still don't know where norms will settle.  The
risk we are concerned with is that lax safeguards in the meantime will
lead to norms that do not accord proper respect to the rights of people
(and we instead get systems that over-optimize for things like reduced
friction).

Closes #5.

* Less em-dash

* Jeffrey's improvements

Co-authored-by: Jeffrey Yasskin <jyasskin@gmail.com>

---------

Co-authored-by: Jeffrey Yasskin <jyasskin@gmail.com>